+++ components/nacelle/ProductCard.vue
@@ -57,7 +57,7 @@
 
 <script>
 import { mapState, mapMutations, mapActions, mapGetters } from 'vuex'
-import getPriceForCurrency from '~/utils/getPriceForCurrency'
+import productPriceForCurrency from '~/utils/productPriceForCurrency'
 
 export default {
   props: {
@@ -133,7 +133,7 @@ export default {
     },
     displayPrice() {
       if (this.selectedVariant) {
-        return getPriceForCurrency({
+        return productPriceForCurrency({
           product: this.product,
           fallbackPrice: this.selectedVariant.price,
           locale: this.locale
