+++ utils/productOptionAvailability.js
